Move TestDialogs into common file
