+++ src/test/java/fi/aalto/cs/apluscourses/intellij/actions/ImportIdeSettingsActionTest.java
@@ -11,7 +11,6 @@ import com.intellij.openapi.project.Project;
 import fi.aalto.cs.apluscourses.model.Course;
 import fi.aalto.cs.apluscourses.presentation.CourseViewModel;
 import fi.aalto.cs.apluscourses.presentation.MainViewModel;
-import fi.aalto.cs.apluscourses.ui.base.Dialogs;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
@@ -19,55 +18,11 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
-import org.jetbrains.annotations.NotNull;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
 public class ImportIdeSettingsActionTest {
-
-  private class TestDialogs implements Dialogs {
-    private boolean answerOk;
-
-    private String lastInformationMessage = "";
-    private String lastErrorMessage = "";
-    private String lastOkCancelMessage = "";
-
-    public TestDialogs(boolean answerOk) {
-      this.answerOk = answerOk;
-    }
-
-    @Override
-    public void showInformationDialog(@NotNull String message, @NotNull String title) {
-      lastInformationMessage = message;
-    }
-
-    @Override
-    public void showErrorDialog(@NotNull String message, @NotNull String title) {
-      lastErrorMessage = message;
-    }
-
-    @Override
-    public boolean showOkCancelDialog(@NotNull String message,
-                                      @NotNull String title,
-                                      @NotNull String okText,
-                                      @NotNull String cancelText) {
-      lastOkCancelMessage = message;
-      return answerOk;
-    }
-
-    public String getLastInformationMessage() {
-      return lastInformationMessage;
-    }
-
-    public String getLastErrorMessage() {
-      return lastErrorMessage;
-    }
-
-    public String getLastOkCancelMessage() {
-      return lastOkCancelMessage;
-    }
-  }
 
   private Project project;
   private MainViewModel mainViewModel;
